refactor: split extension into per-feature entry points
- provider/ - Synthetic model provider
- web-search/ - Zero-data-retention web search tool
- command-quotas/ - API usage quotas command

Enables users to toggle features individually via pi config.
