+++ package.json
@@ -15,7 +15,9 @@
   },
   "pi": {
     "extensions": [
-      "./src/index.ts"
+      "./src/extensions/provider/index.ts",
+      "./src/extensions/web-search/index.ts",
+      "./src/extensions/command-quotas/index.ts"
     ],
     "video": "https://assets.aliou.me/pi-extensions/demos/pi-synthetic.mp4"
   },
