+++ src/extensions/command-quotas/components/tabbed-panel.ts
