+++ src/lib/env.ts
@@ -1,0 +1,7 @@
+export function hasSyntheticApiKey(): boolean {
+  return Boolean(process.env.SYNTHETIC_API_KEY);
+}
+
+export function getSyntheticApiKey(): string {
+  return process.env.SYNTHETIC_API_KEY || "";
+}
